prepare language files
