--- src/data/landing/es.ts
@@ -1,27 +1,0 @@
-import { InstanceLandingData } from '@/data-types'
-
-export const esInstanceLandingData: InstanceLandingData = {
-  lang: 'es',
-  strings: {
-    vision:
-      'Somos una organización de base comunitaria sin ánimo de lucro que apoya el aprendizaje personalizado y trabajando por la igualdad de oportunidades educativas. Esta plataforma de aprendizaje ofrece miles de artículos de instrucción, videos de aprendizaje y ejercicios de práctica para millones de estudiantes en todo el mundo - completamente gratis. Ahora es el momento de unirse a nosotros en tu idioma.',
-    learnMore: 'Aprende más',
-    democraticallyStructured: 'estructurado democráticamente',
-    nonProfit: 'sin ánimo de lucro',
-    transparent: 'transparente',
-    openlyLicensed: 'con licencia abierta',
-    adFree: 'sin publicidad',
-    freeOfCharge: 'gratuito',
-    wikiTitle: 'Serlo es la Wikipedia para el Aprendizaje',
-    wikiText:
-      'Al igual que Wikipedia, esta plataforma es creada por una comunidad de autores comprometidos con la educación. “Serlo Education” es propiedad de varios equipos descentralizados de voluntarios y profesionales, y manejado por ellos mismos alrededor del mundo.',
-    movementTitle:
-      'Conviértete en Parte de Nuestro Movimiento para la Educación Abierta',
-    callForAuthors:
-      'Estamos buscando profesores, maestros y educadores entusiastas a quienes les apasionen las materias que enseñan. ¡Conviértete en un autor en serlo.org! Tu puedes crear nuevo material de aprendizaje y ayudarnos a mejorar el contenido existente.',
-    communityLink: 'Visita la página para autores',
-    callForOther:
-      'Nosotros ofrecemos una diversa gama de trabajo y oportunidades de voluntariado, dentro del campo del desarrollo de software, diseño, traducción, comunicación, administración de proyectos, y más.',
-    getInvolved: '¡Participa!',
-  },
-}
