--- src/data/landing/fr.ts
@@ -1,26 +1,0 @@
-import { InstanceLandingData } from '@/data-types'
-
-export const frInstanceLandingData: InstanceLandingData = {
-  lang: 'fr',
-  strings: {
-    vision:
-      "Nous sommes une organisation communautaire à but non lucratif supportant l'apprentissage personnalisé et travaillant à la réalisation d'opportunités éducatives égales. Cette plateforme propose des milliers d'articles d'instruction, des vidéos pédagogiques et des exercices pratiques aux millions d'étudiant(e)s et d'élèves dans le monde entier - complètement gratuit. Maintenant, vous êtes invité de joindre l'équipe Serlo francophone.",
-    learnMore: 'En savoir plus',
-    democraticallyStructured: 'structure démocratique',
-    nonProfit: 'non-lucratif',
-    transparent: 'transparent',
-    openlyLicensed: 'licence libre',
-    adFree: 'sans publicité',
-    freeOfCharge: 'gratuit',
-    wikiTitle: "Serlo est le Wikipédia pour l'apprentissage",
-    wikiText:
-      "Tout comme Wikipédia, cette plateforme est créée par une communauté d'auteurs engagé(e)s. Serlo Education est gérée et détenue par des équipes décentralisées de bénévoles et de professionnels dans le monde entier.",
-    movementTitle: "Rejoignez notre mouvement pour l'éducation libre",
-    callForAuthors:
-      'Nous cherchons des enseignant(e)s et des éducateur(e)s enthousiastes et passionné(e)s de leur matière. Devenez un(e) auteur sur serlo.org ! Vous pouvez créer du nouveau matériel pédagogique et nous aider à améliorer le contenu existant.',
-    communityLink: "Visitez la page d'accueil d'auteurs",
-    callForOther:
-      "Nous offrons une variété d'emplois et de possibilités de bénévolat dans les domaines du développement de logiciel, de la conception, la traduction, la communication, la gestion de projet et d'autres.",
-    getInvolved: 'Participe!',
-  },
-}
