--- src/data/landing/hi.ts
@@ -1,26 +1,0 @@
-import { InstanceLandingData } from '@/data-types'
-
-export const hiInstanceLandingData: InstanceLandingData = {
-  lang: 'hi',
-  strings: {
-    vision:
-      'हम नॉनप्रॉफिट आर्गेनाइजेशन संगठन हैं जो व्यक्तिगत शिक्षा का समर्थन करते हैं और समान शैक्षणिक अवसरों के लिए काम करते हैं। यह सीखने का मंच दुनिया भर के लाखों छात्रों के लिए हजारों निर्देशक लेख, वीडियो के माध्यम से सीखना और अभ्यास प्रदान करता है - पूरी तरह से नि: शुल्क। अब हिंदी में हमसे जुड़ने का समय है।',
-    learnMore: 'और जानो',
-    democraticallyStructured: 'लोकतांत्रिक',
-    nonProfit: 'नॉनप्रॉफिट',
-    transparent: 'पारदर्शक',
-    openlyLicensed: 'खुले तौर पर लाइसेंस प्राप्त',
-    adFree: 'विज्ञापन मुक्त',
-    freeOfCharge: 'निःशुल्क',
-    wikiTitle: 'सेर्लो सीखने के लिए विकिपीडिया है।',
-    wikiText:
-      'विकिपीडिया की तरह, यह मंच लेखकों के एक व्यस्त समुदाय द्वारा बनाया गया है। सेर्लो शिक्षा पूरी दुनिया में स्वयंसेवकों और पेशेवरों की विकेन्द्रीकृत टीमों द्वारा संचालित और स्वामित्व में है।',
-    movementTitle: 'ओपन एजुकेशन के लिए हमारे आंदोलन का एक हिस्सा बनें',
-    callForAuthors:
-      'हम उत्साही शिक्षकों की तलाश में हैं जो अपने विषय के बारे में भावुक हैं। Serlo.org पर एक लेखक बनें! आप नई सीखने की सामग्री बना सकते हैं और मौजूदा सामग्री को बेहतर बनाने में हमारी सहायता कर सकते हैं।',
-    communityLink: 'लेखकों के लिए लैंडिंग पृष्ठ पर जाएं',
-    callForOther:
-      'हम सॉफ्टवेयर विकास, डिजाइन, अनुवाद, संचार, परियोजना प्रबंधन और अन्य क्षेत्रों के क्षेत्र में नौकरियों और स्वयंसेवी अवसरों की विविध श्रेणी प्रदान करते हैं।',
-    getInvolved: 'सेर्लो में शामिल हो जाएं',
-  },
-}
