--- src/data/landing/ta.ts
@@ -1,26 +1,0 @@
-import { InstanceLandingData } from '@/data-types'
-
-export const taInstanceLandingData: InstanceLandingData = {
-  lang: 'ta',
-  strings: {
-    vision:
-      'நாம் சமமான கல்வி வாய்ப்புகளை நோக்கி இணைந்து பணிபுரியும் ஒரு குழு. இந்த இணையத்தளத்தில் எண்ணற்ற விவரக் கட்டுரைகள், பயிற்சிகள் மற்றும் ஒலிப் பேழைகள் அனைத்துப் பாடங்களுக்கும் வழங்கப்பட்டுவருகின்றன. இவை அனைத்தும் இலவசமாக உலகம் முழுவதும் உள்ள மாணவர்களுக்காக உருவாக்கப்பட்டுவருகின்றன. இனி வரும் காலங்களில், தமிழ்மொழியிலும் இவ்வாறான இலவசப் பாடத்திட்டங்களை உருவாக்க நீங்களும் எம்முடன் இணைந்து பணியாற்றலாம்.',
-    learnMore: 'மேலும் அறிக',
-    democraticallyStructured: 'ஜனநாயக ரீதியாக',
-    nonProfit: 'இலாப நோக்கற்றது',
-    transparent: 'ஒளி புகும்',
-    openlyLicensed: 'திறந்த உரிமம்',
-    adFree: 'விளம்பரமின்றி',
-    freeOfCharge: 'இலவசம்',
-    wikiTitle: 'Serlo ஓர் கற்றலுக்கான விக்கிபீடியா',
-    wikiText:
-      'Serlo.org விக்கிபீடியாபோல திறந்த உரிமம் கொண்ட ஓர் இணையத்தளம். இது எம் எழுத்தாளர் குழுவால் உருவாக்கப்படுகின்றது.',
-    movementTitle: 'நீங்களும் இதில் பணியாற்றலாம்',
-    callForAuthors:
-      'ஆசிரியர்களும் ஆர்வமுள்ள எழுத்தாளர்களும் பாடங்களை உருவாக்க பல வழிகளில் உதவலாம். புதுப் பயிற்சிகளை உருவாக்குவதற்கும் இந்தத் தளத்தின் சில உள்ளடக்கங்களை இன்னும் மேம்படுத்துவதற்கும் நீங்கள் உதவலாம். அதற்கு கீழுள்ள இணையத்திற்குச் செல்லவும்.',
-    communityLink: 'எழுத்தாளருக்கான பக்கத்தைப் பார்வையிடவும்',
-    callForOther:
-      'நாங்கள் பல வகையான வேலைவாய்ப்புகளையும் பொதுச்சேவையாகப் பணியாற்றும் வாய்ப்புகளையும் வழங்குகின்றோம். இந்த இணையத்தளதிற்கு மொழிபெயர்ப்பாளர்கள், வடிவமைப்பாளர்கள், தொலைத்தொடர்பாளர்கள் போன்ற துறை சார்ந்தவர்களை நாங்கள் தேடி நிற்கின்றோம். இணைந்து கொள்ளுங்கள்.',
-    getInvolved: 'நீங்களும் ஈடுபடுங்கள்!',
-  },
-}
